fix: remove unnecessary event subscription and reorganize imports
